+++ src/features/auth/hooks/useRegisterForm.ts
@@ -3,10 +3,11 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation } from '@tanstack/react-query';
 
-import { RegisterFormSchema } from '@schemas/auth/register.schema';
-import type { RegisterForm } from '@schemas/auth/register.schema';
 import { signUp } from '@server/auth/signUp';
 import type { FormNotification } from '@typings/forms/notification';
+
+import { RegisterFormSchema } from '../schemas/register.schema';
+import type { RegisterForm } from '../schemas/register.schema';
 
 export default function useRegisterForm() {
   const [registerFormNotification, setRegisterFormNotification] = useState<FormNotification>({