+++ src/server/auth/signIn.ts
@@ -1,5 +1,6 @@
 import { supabase } from '@lib/supabase/client';
-import type { LoginForm } from '@schemas/auth/login.schema';
+
+import type { LoginForm } from '@features/auth/schemas/login.schema';
 import { createLogEvent } from '@lib/supabase/logEvent';
 
 export async function signIn(formData: LoginForm) {